requirements.txt
google-auth-oauthlib
google-api-python-client
flask
gunicorn
